# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GTrader API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /psychology:
    get:
      operationId: listPsychologyLogs
      tags: [psychology]
      summary: List psychology logs
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of psychology logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PsychologyLog"
    post:
      operationId: createPsychologyLog
      tags: [psychology]
      summary: Create a psychology log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePsychologyLogBody"
      responses:
        "201":
          description: Created psychology log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsychologyLog"
  /psychology/{id}:
    patch:
      operationId: updatePsychologyLog
      tags: [psychology]
      summary: Update a psychology log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePsychologyLogBody"
      responses:
        "200":
          description: Updated psychology log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsychologyLog"
    delete:
      operationId: deletePsychologyLog
      tags: [psychology]
      summary: Delete a psychology log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /analytics/timeframes:
    get:
      operationId: getTimeframeAnalytics
      tags: [analytics]
      summary: Get win rates, RR, and profitability broken down by timeframe
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Timeframe analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeframeStats"
  /analytics/triggers:
    get:
      operationId: getTriggerAnalytics
      tags: [analytics]
      summary: Get performance by entry trigger type
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Trigger analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TriggerStats"
  /analytics/sessions:
    get:
      operationId: getSessionAnalytics
      tags: [analytics]
      summary: Get performance by trading session
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Session analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionStats"
  /analytics/setups:
    get:
      operationId: getSetupAnalytics
      tags: [analytics]
      summary: Get performance by setup type
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Setup analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SetupStats"
  /analytics/equity-curve:
    get:
      operationId: getEquityCurve
      tags: [analytics]
      summary: Get equity curve data points for charting
      parameters:
        - name: accountId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equity curve data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquityPoint"
  /analytics/recent-trades:
    get:
      operationId: getRecentTrades
      tags: [analytics]
      summary: Get most recent trades across all accounts
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recent trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
  /analytics/combinations:
    get:
      operationId: getCombinationAnalytics
      tags: [analytics]
      summary: Best performing timeframe+trigger+setup combinations
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Combination analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CombinationStats"
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        tradeDate:
          type: string
        asset:
          type: string
        direction:
          type: string
          description: "buy, sell"
        strategy:
          type: ["string", "null"]
        setup:
          type: ["string", "null"]
        timeframe:
          type: ["string", "null"]
          description: "M1, M2, M3, M5, M10, M15, M30, H1"
        higherTimeframeBias:
          type: ["string", "null"]
          description: "H1, H4, Daily"
        entryTrigger:
          type: ["string", "null"]
          description: "MSS, BOS, LiquiditySweep, OrderBlock, FVG, SMTDivergence, RejectionCandle, Engulfing, BreakerBlock, MitigationBlock, SessionManipulation, CHOCH"
        session:
          type: ["string", "null"]
          description: "london, new_york, asian, london_open, ny_open"
        entryPrice:
          type: ["number", "null"]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        riskPercent:
          type: ["number", "null"]
        riskAmount:
          type: ["number", "null"]
        riskReward:
          type: ["number", "null"]
        result:
          type: string
          description: "win, loss, breakeven"
        pnl:
          type: ["number", "null"]
        followedPlan:
          type: ["boolean", "null"]
        validSetup:
          type: ["boolean", "null"]
        impulsiveTrade:
          type: ["boolean", "null"]
        riskRespected:
          type: ["boolean", "null"]
        emotionBefore:
          type: ["string", "null"]
          description: "confident, fearful, anxious, overconfident, tired, revenge, focused"
        emotionAfter:
          type: ["string", "null"]
          description: "calm, frustrated, disciplined, emotional, happy, impulsive"
        disciplineScore:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - accountId
        - tradeDate
        - asset
        - direction
        - result
        - createdAt
        - updatedAt
...
    PsychologyLog:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: ["integer", "null"]
        logDate:
          type: string
        emotionState:
          type: string
        energyLevel:
          type: ["integer", "null"]
        stressLevel:
          type: ["integer", "null"]
        focusLevel:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tradedToday:
          type: ["boolean", "null"]
        followedRules:
          type: ["boolean", "null"]
        overallScore:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - logDate
        - emotionState
        - createdAt
    CreatePsychologyLogBody:
      type: object
      properties:
        accountId:
          type: ["integer", "null"]
        logDate:
          type: string
        emotionState:
          type: string
        energyLevel:
          type: ["integer", "null"]
        stressLevel:
          type: ["integer", "null"]
        focusLevel:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tradedToday:
          type: ["boolean", "null"]
        followedRules:
          type: ["boolean", "null"]
        overallScore:
          type: ["integer", "null"]
      required:
        - logDate
        - emotionState
    UpdatePsychologyLogBody:
      type: object
      properties:
        emotionState:
          type: ["string", "null"]
        energyLevel:
          type: ["integer", "null"]
        stressLevel:
          type: ["integer", "null"]
        focusLevel:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tradedToday:
          type: ["boolean", "null"]
        followedRules:
          type: ["boolean", "null"]
        overallScore:
          type: ["integer", "null"]
...
    TimeframeStats:
      type: object
      properties:
        timeframe:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        averageRR:
          type: number
        averageWin:
          type: number
        averageLoss:
          type: number
        profitFactor:
          type: number
      required:
        - timeframe
        - totalTrades
        - wins
        - losses
        - winRate
        - totalPnl
        - averageRR
        - averageWin
        - averageLoss
        - profitFactor
    TriggerStats:
      type: object
      properties:
        entryTrigger:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        averageRR:
          type: number
        profitFactor:
          type: number
      required:
        - entryTrigger
        - totalTrades
        - wins
        - losses
        - winRate
        - totalPnl
        - averageRR
        - profitFactor
    SessionStats:
      type: object
      properties:
        session:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        averageRR:
          type: number
        profitFactor:
          type: number
      required:
        - session
        - totalTrades
        - wins
        - losses
        - winRate
        - totalPnl
        - averageRR
        - profitFactor
    SetupStats:
      type: object
      properties:
        setup:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        averageRR:
          type: number
        profitFactor:
          type: number
      required:
        - setup
        - totalTrades
        - wins
        - losses
        - winRate
        - totalPnl
        - averageRR
        - profitFactor
    EquityPoint:
      type: object
      properties:
        date:
          type: string
        balance:
          type: number
        pnl:
          type: number
        tradeId:
          type: ["integer", "null"]
      required:
        - date
        - balance
        - pnl
    CombinationStats:
      type: object
      properties:
        timeframe:
          type: ["string", "null"]
        entryTrigger:
          type: ["string", "null"]
        setup:
          type: ["string", "null"]
        totalTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        averageRR:
          type: number
        profitFactor:
          type: number
      required:
        - totalTrades
        - winRate
        - totalPnl
        - averageRR
        - profitFactor